// inclusive-city/src/app/api/overpassobjectApi.ts
export interface ReviewDto {
    id?: number;
    osmId?: number;
    osmType?: string;
    comment?: string;
    photoUrl?: string | null;
    createdBy?: string;
    username?: string;
    createdAt?: Date;
    rate?: number;
}

export interface AddReviewCommand {
    osmId?: number;
    osmType?: string;
    comment?: string;
    imageBase64?: string | null;
    createdBy?: string;
    username?: string;
    rate?: number;
}
// ...
export interface GetStructuresDto {
    elements?: ElementDto[];
}

export interface OverpassElementDto {
    id?: number;
    type?: string;
    lat?: number | null;
    lon?: number | null;
    nodes?: number[] | null;
    tags?: { [key: string]: string; } | null;
    members?: OverpassMemberDto[] | null;
}

export interface ElementDto extends OverpassElementDto {
    rating?: number;
    reviews?: ReviewDto[] | null;
    imageUrls?: string[] | null;
}

export interface OverpassMemberDto {
    type?: string;
    ref?: number;
    role?: string;
}

export interface UploadStructureImagesCommand {
    osmId?: number;
    images?: ImageDto[];
}

export interface ImageDto {
    imageBase64?: string | null;
}

export interface FileResponse {
    data: Blob;
    status: number;
    fileName?: string;
    headers?: { [name: string]: any };
}